+++ apps/frontend/src/app/page.tsx
@@ -1,5 +1,7 @@
 import Image from "next/image";
 
+// Home page
+// TODO: Make this a real home page
 export default function Home() {
   return (<div><div>Home</div></div>
   );
