+++ apps/frontend/src/components/resume/editor/fragments/ec.fragment.tsx
@@ -1,8 +1,8 @@
-import PlusButton from "../../plusbutton";
+import PlusButton from "../../../plusbutton";
 
-export default function ExperienceFragment() {
+export default function ECFragment() {
     return (<div className="text-xl font-bold flex flex-row">
-        <div>Experience</div>
+        <div>Extra Curriculars</div>
         <div className="flex-1"></div>
         <PlusButton />
     </div>)
