+++ apps/frontend/src/components/resume/pdf/resume.tsx
@@ -8,6 +8,8 @@ const libraries = ['Google Cloud Platform', 'Pandas', 'TensorFlow', 'SciKit Lear
 const hobbies = ['Sketching', 'Painting', 'Violin (Western and Carnatic)', 'Badminton', 'Swimming', 'Chess']
 const certifications = ['AWS Cloud Practitioner', 'Docker & Kubernetes']
 
+// Resume details - hardcoded for now
+// This will be replaced with a database call in the future
 export const resume: namespace.ResumeDetails = {
     name: "Aditya Rao",
     contactInfo: {
