+++ apps/frontend/src/state/redux-provider.tsx
@@ -1,6 +1,7 @@
 import { Provider } from "react-redux";
 import { store } from "@/state/store";
 
+// Provider for the redux store
 export default function ReduxProvider({
     children,
 }: {
